--- backend/utils/preprocess.py
@@ -1,11 +1,0 @@
-def clean_prediction(decoded_text):
-    special_tokens = [
-        "<|en|>",
-        "<|transcribe|>",
-        "<|notimestamps|>",
-        "<|endoftext|>",
-        "<|startoftranscript|>",
-    ]
-    for token in special_tokens:
-        decoded_text = decoded_text.replace(token, "").strip()
-    return decoded_text
